Initial Commit
An experimental conversion of the GMod 10-era GLua highlighting plugin. It is not perfect, as the old version is so old it had to essentially be totally converted from scratch. Please report bugs and make merge requests if you know how to fix any of the specific quirks with it.